requirements.txt
streamlit==1.40.1
numpy==1.26.4
tensorflow==2.17.1
Pillow==10.4.0
